[common] Fix SoundcloudItem.id type and hash calculation for large IDs
Changed SoundcloudItem.id from int to long to support SoundCloud's large ID
values that exceed Integer.MAX_VALUE. Updated SoundcloudSearchResult hash
calculation to use Long.toHexString() instead of Integer.toHexString().

This fixes JSON parsing errors when deserializing search results from the
SoundCloud API where track IDs can be larger than 2.1 billion.

+++ common/src/main/java/com/frostwire/search/soundcloud/SoundcloudSearchResult.java
@@ -66,7 +66,7 @@ public final class SoundcloudSearchResult extends AbstractFileSearchResult imple
         } else {
             this.progressiveFormatJSONFetcherURL = null;
         }
-        this.hash = Integer.toHexString(item.id * 953 * 631);
+        this.hash = Long.toHexString(item.id * 953 * 631);
         this.downloadUrl = null;
     }
